requirements.txt
fastapi
uvicorn
requests
apscheduler
psycopg2-binary
sqlalchemy
python-dotenv

